@@ -21,16 +21,15 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"reflect"
 	"strconv"
 	"strings"
 	"testing"
 
-	k8gbv1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
-	"github.com/k8gb-io/k8gb/controllers/internal/utils"
+	utils2 "github.com/k8gb-io/k8gb/controllers/utils"
 
+	k8gbv1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
 	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/assert"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -56,10 +55,11 @@ const (
 
 var predefinedConfig = Config{
 	ReconcileRequeueSeconds: 30,
+	NSRecordTTL:             30,
 	ClusterGeoTag:           "us",
 	ExtClustersGeoTags:      []string{"za", "eu"},
 	EdgeDNSType:             DNSTypeInfoblox,
-	EdgeDNSServers: []utils.DNSServer{
+	EdgeDNSServers: []utils2.DNSServer{
 		{
 			Host: "dns.cloud.example.com",
 			Port: 53,
@@ -265,6 +265,24 @@ func TestResolveConfigWithZeroReconcileRequeueSecondsKey(t *testing.T) {
 	arrangeVariablesAndAssert(t, expected, assert.Error)
 }
 
+func TestResolveConfigWithNegativeNSRecordTTL(t *testing.T) {
+	// arrange
+	defer cleanup()
+	expected := predefinedConfig
+	expected.NSRecordTTL = -1
+	// act,assert
+	arrangeVariablesAndAssert(t, expected, assert.Error)
+}
+
+func TestResolveConfigWithZeroReconcileNSRecordTTL(t *testing.T) {
+	// arrange
+	defer cleanup()
+	expected := predefinedConfig
+	expected.NSRecordTTL = 0
+	// act,assert
+	arrangeVariablesAndAssert(t, expected, assert.Error)
+}
+
 // remove this test once the deprecated key is no longer supported
 func TestResolveConfigWithDeprecatedEdgeDNSServerKey(t *testing.T) {
 	// arrange
@@ -349,7 +367,7 @@ func TestResolveConfigWithEmptyEdgeDnsServer(t *testing.T) {
 	// arrange
 	defer cleanup()
 	expected := predefinedConfig
-	expected.EdgeDNSServers = []utils.DNSServer{}
+	expected.EdgeDNSServers = []utils2.DNSServer{}
 	// act,assert
 	arrangeVariablesAndAssert(t, expected, assert.Error)
 }
@@ -358,7 +376,7 @@ func TestResolveConfigWithTwoEdgeDnsServers(t *testing.T) {
 	// arrange
 	defer cleanup()
 	expected := predefinedConfig
-	expected.EdgeDNSServers = []utils.DNSServer{
+	expected.EdgeDNSServers = []utils2.DNSServer{
 		{
 			Host: "8.8.8.8",
 			Port: 53,
@@ -376,7 +394,7 @@ func TestResolveConfigWithNoEdgeDnsServer(t *testing.T) {
 	// arrange
 	defer cleanup()
 	expected := predefinedConfig
-	expected.EdgeDNSServers = []utils.DNSServer{}
+	expected.EdgeDNSServers = []utils2.DNSServer{}
 	// act,assert
 	arrangeVariablesAndAssert(t, expected, assert.Error, EdgeDNSServersKey)
 }
@@ -385,7 +403,7 @@ func TestResolveConfigWithEmptyIpAddressInEdgeDnsServer(t *testing.T) {
 	// arrange
 	defer cleanup()
 	expected := predefinedConfig
-	expected.EdgeDNSServers = []utils.DNSServer{
+	expected.EdgeDNSServers = []utils2.DNSServer{
 		{
 			Host: defaultEdgeDNSServerIP,
 			Port: 53,
@@ -399,7 +417,7 @@ func TestResolveConfigWithHostnameEdgeDnsServer(t *testing.T) {
 	// arrange
 	defer cleanup()
 	expected := predefinedConfig
-	expected.EdgeDNSServers = []utils.DNSServer{
+	expected.EdgeDNSServers = []utils2.DNSServer{
 		{
 			Host: "server-nonprod.on.domain.l3.2l.com",
 			Port: 53,
@@ -414,7 +432,7 @@ func TestResolveConfigWithInvalidIpAddressEdgeDnsServer(t *testing.T) {
 	// arrange
 	defer cleanup()
 	expected := predefinedConfig
-	expected.EdgeDNSServers = []utils.DNSServer{
+	expected.EdgeDNSServers = []utils2.DNSServer{
 		{
 			Host: fmt.Sprintf("%s.", defaultEdgeDNSServerIP),
 			Port: 53,
@@ -1252,7 +1270,7 @@ func TestNsServerNamesForLocalEdgeDNS(t *testing.T) {
 	defer cleanup()
 	for _, edgeDNSServer := range []string{"127.0.0.1", "localhost"} {
 		customConfig := predefinedConfig
-		customConfig.EdgeDNSServers = []utils.DNSServer{
+		customConfig.EdgeDNSServers = []utils2.DNSServer{
 			{
 				Host: edgeDNSServer,
 				Port: 53,
@@ -1505,8 +1523,8 @@ func arrangeVariablesAndAssert(t *testing.T, expected Config,
 }
 
 func cleanup() {
-	for _, s := range []string{ReconcileRequeueSecondsKey, ClusterGeoTagKey, ExtClustersGeoTagsKey, EdgeDNSZoneKey, DNSZoneKey, EdgeDNSServersKey,
-		ExtDNSEnabledKey, InfobloxGridHostKey, InfobloxVersionKey, InfobloxPortKey, InfobloxUsernameKey,
+	for _, s := range []string{ReconcileRequeueSecondsKey, NSRecordTTLKey, ClusterGeoTagKey, ExtClustersGeoTagsKey, EdgeDNSZoneKey, DNSZoneKey,
+		EdgeDNSServersKey, ExtDNSEnabledKey, InfobloxGridHostKey, InfobloxVersionKey, InfobloxPortKey, InfobloxUsernameKey,
 		InfobloxPasswordKey, K8gbNamespaceKey, CoreDNSExposedKey, InfobloxHTTPRequestTimeoutKey,
 		InfobloxHTTPPoolConnectionsKey, LogLevelKey, LogFormatKey, LogNoColorKey, MetricsAddressKey, SplitBrainCheckKey, TracingEnabled,
 		TracingSamplingRatio, OtelExporterOtlpEndpoint} {
@@ -1520,6 +1538,7 @@ func configureEnvVar(config Config) {
 	_ = os.Unsetenv(EdgeDNSServerKey)
 	_ = os.Unsetenv(EdgeDNSServerPortKey)
 	_ = os.Setenv(ReconcileRequeueSecondsKey, strconv.Itoa(config.ReconcileRequeueSeconds))
+	_ = os.Setenv(NSRecordTTLKey, strconv.Itoa(config.NSRecordTTL))
 	_ = os.Setenv(ClusterGeoTagKey, config.ClusterGeoTag)
 	_ = os.Setenv(ExtClustersGeoTagsKey, strings.Join(config.ExtClustersGeoTags, ","))
 	_ = os.Setenv(EdgeDNSServersKey, config.EdgeDNSServers.String())
@@ -1547,11 +1566,11 @@ func configureEnvVar(config Config) {
 
 func getTestContext(testData string) (client.Client, *k8gbv1beta1.Gslb) {
 	// Create a fake client to mock API calls.
-	var gslbYaml, err = ioutil.ReadFile(testData)
+	var gslbYaml, err = os.ReadFile(testData)
 	if err != nil {
 		panic(fmt.Errorf("can't open example CR file: %s", testData))
 	}
-	gslb, err := utils.YamlToGslb(gslbYaml)
+	gslb, err := utils2.YamlToGslb(gslbYaml)
 	if err != nil {
 		panic(err)
 	}
