@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: controllers/providers/assistant/assistant.go
+//
+// Generated by this command:
+//
+//	mockgen -package=mocks -destination=controllers/mocks/assistant_mock.go -source=controllers/providers/assistant/assistant.go Assistant
+//
 
 // Package mocks is a generated GoMock package.
 package mocks
@@ -26,9 +31,9 @@ import (
 	reflect "reflect"
 	time "time"
 
-	gomock "github.com/golang/mock/gomock"
-	v1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
 	assistant "github.com/k8gb-io/k8gb/controllers/providers/assistant"
+	gomock "go.uber.org/mock/gomock"
+	v1 "k8s.io/api/core/v1"
 	endpoint "sigs.k8s.io/external-dns/endpoint"
 )
 
@@ -70,6 +75,21 @@ func (mr *MockAssistantMockRecorder) CoreDNSExposedIPs() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CoreDNSExposedIPs", reflect.TypeOf((*MockAssistant)(nil).CoreDNSExposedIPs))
 }
 
+// GetCoreDNSService mocks base method.
+func (m *MockAssistant) GetCoreDNSService() (*v1.Service, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "GetCoreDNSService")
+	ret0, _ := ret[0].(*v1.Service)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// GetCoreDNSService indicates an expected call of GetCoreDNSService.
+func (mr *MockAssistantMockRecorder) GetCoreDNSService() *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetCoreDNSService", reflect.TypeOf((*MockAssistant)(nil).GetCoreDNSService))
+}
+
 // GetExternalTargets mocks base method.
 func (m *MockAssistant) GetExternalTargets(host string, extClusterNsNames map[string]string) assistant.Targets {
 	m.ctrl.T.Helper()
@@ -79,24 +99,9 @@ func (m *MockAssistant) GetExternalTargets(host string, extClusterNsNames map[st
 }
 
 // GetExternalTargets indicates an expected call of GetExternalTargets.
-func (mr *MockAssistantMockRecorder) GetExternalTargets(host, extClusterNsNames interface{}) *gomock.Call {
+func (mr *MockAssistantMockRecorder) GetExternalTargets(host, extClusterNsNames any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetExternalTargets", reflect.TypeOf((*MockAssistant)(nil).GetExternalTargets), host, extClusterNsNames)
-}
-
-// GslbIngressExposedIPs mocks base method.
-func (m *MockAssistant) GslbIngressExposedIPs(gslb *v1beta1.Gslb) ([]string, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GslbIngressExposedIPs", gslb)
-	ret0, _ := ret[0].([]string)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// GslbIngressExposedIPs indicates an expected call of GslbIngressExposedIPs.
-func (mr *MockAssistantMockRecorder) GslbIngressExposedIPs(gslb interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GslbIngressExposedIPs", reflect.TypeOf((*MockAssistant)(nil).GslbIngressExposedIPs), gslb)
 }
 
 // InspectTXTThreshold mocks base method.
@@ -108,7 +113,7 @@ func (m *MockAssistant) InspectTXTThreshold(fqdn string, splitBrainThreshold tim
 }
 
 // InspectTXTThreshold indicates an expected call of InspectTXTThreshold.
-func (mr *MockAssistantMockRecorder) InspectTXTThreshold(fqdn, splitBrainThreshold interface{}) *gomock.Call {
+func (mr *MockAssistantMockRecorder) InspectTXTThreshold(fqdn, splitBrainThreshold any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "InspectTXTThreshold", reflect.TypeOf((*MockAssistant)(nil).InspectTXTThreshold), fqdn, splitBrainThreshold)
 }
@@ -122,7 +127,7 @@ func (m *MockAssistant) RemoveEndpoint(endpointName string) error {
 }
 
 // RemoveEndpoint indicates an expected call of RemoveEndpoint.
-func (mr *MockAssistantMockRecorder) RemoveEndpoint(endpointName interface{}) *gomock.Call {
+func (mr *MockAssistantMockRecorder) RemoveEndpoint(endpointName any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RemoveEndpoint", reflect.TypeOf((*MockAssistant)(nil).RemoveEndpoint), endpointName)
 }
@@ -136,7 +141,7 @@ func (m *MockAssistant) SaveDNSEndpoint(namespace string, i *endpoint.DNSEndpoin
 }
 
 // SaveDNSEndpoint indicates an expected call of SaveDNSEndpoint.
-func (mr *MockAssistantMockRecorder) SaveDNSEndpoint(namespace, i interface{}) *gomock.Call {
+func (mr *MockAssistantMockRecorder) SaveDNSEndpoint(namespace, i any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SaveDNSEndpoint", reflect.TypeOf((*MockAssistant)(nil).SaveDNSEndpoint), namespace, i)
 }
