@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/infobloxopen/infoblox-go-client (interfaces: IBConnector)
+//
+// Generated by this command:
+//
+//	mockgen -package=mocks -destination=controllers/mocks/infoblox-connection_mock.go github.com/infobloxopen/infoblox-go-client IBConnector
+//
 
 // Package mocks is a generated GoMock package.
 package mocks
@@ -25,8 +30,8 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	ibclient "github.com/infobloxopen/infoblox-go-client"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockIBConnector is a mock of IBConnector interface.
@@ -62,7 +67,7 @@ func (m *MockIBConnector) CreateObject(arg0 ibclient.IBObject) (string, error) {
 }
 
 // CreateObject indicates an expected call of CreateObject.
-func (mr *MockIBConnectorMockRecorder) CreateObject(arg0 interface{}) *gomock.Call {
+func (mr *MockIBConnectorMockRecorder) CreateObject(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateObject", reflect.TypeOf((*MockIBConnector)(nil).CreateObject), arg0)
 }
@@ -77,13 +82,13 @@ func (m *MockIBConnector) DeleteObject(arg0 string) (string, error) {
 }
 
 // DeleteObject indicates an expected call of DeleteObject.
-func (mr *MockIBConnectorMockRecorder) DeleteObject(arg0 interface{}) *gomock.Call {
+func (mr *MockIBConnectorMockRecorder) DeleteObject(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeleteObject", reflect.TypeOf((*MockIBConnector)(nil).DeleteObject), arg0)
 }
 
 // GetObject mocks base method.
-func (m *MockIBConnector) GetObject(arg0 ibclient.IBObject, arg1 string, arg2 interface{}) error {
+func (m *MockIBConnector) GetObject(arg0 ibclient.IBObject, arg1 string, arg2 any) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetObject", arg0, arg1, arg2)
 	ret0, _ := ret[0].(error)
@@ -91,7 +96,7 @@ func (m *MockIBConnector) GetObject(arg0 ibclient.IBObject, arg1 string, arg2 in
 }
 
 // GetObject indicates an expected call of GetObject.
-func (mr *MockIBConnectorMockRecorder) GetObject(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockIBConnectorMockRecorder) GetObject(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetObject", reflect.TypeOf((*MockIBConnector)(nil).GetObject), arg0, arg1, arg2)
 }
@@ -106,7 +111,7 @@ func (m *MockIBConnector) UpdateObject(arg0 ibclient.IBObject, arg1 string) (str
 }
 
 // UpdateObject indicates an expected call of UpdateObject.
-func (mr *MockIBConnectorMockRecorder) UpdateObject(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockIBConnectorMockRecorder) UpdateObject(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpdateObject", reflect.TypeOf((*MockIBConnector)(nil).UpdateObject), arg0, arg1)
 }
