@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: sigs.k8s.io/controller-runtime/pkg/manager (interfaces: Manager)
+//
+// Generated by this command:
+//
+//	mockgen -package=mocks -destination=controllers/mocks/manager_mock.go sigs.k8s.io/controller-runtime/pkg/manager Manager
+//
 
 // Package mocks is a generated GoMock package.
 package mocks
@@ -28,14 +33,14 @@ import (
 	reflect "reflect"
 
 	logr "github.com/go-logr/logr"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 	meta "k8s.io/apimachinery/pkg/api/meta"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 	rest "k8s.io/client-go/rest"
 	record "k8s.io/client-go/tools/record"
 	cache "sigs.k8s.io/controller-runtime/pkg/cache"
 	client "sigs.k8s.io/controller-runtime/pkg/client"
-	v1alpha1 "sigs.k8s.io/controller-runtime/pkg/config/v1alpha1"
+	config "sigs.k8s.io/controller-runtime/pkg/config"
 	healthz "sigs.k8s.io/controller-runtime/pkg/healthz"
 	manager "sigs.k8s.io/controller-runtime/pkg/manager"
 	webhook "sigs.k8s.io/controller-runtime/pkg/webhook"
@@ -73,7 +78,7 @@ func (m *MockManager) Add(arg0 manager.Runnable) error {
 }
 
 // Add indicates an expected call of Add.
-func (mr *MockManagerMockRecorder) Add(arg0 interface{}) *gomock.Call {
+func (mr *MockManagerMockRecorder) Add(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Add", reflect.TypeOf((*MockManager)(nil).Add), arg0)
 }
@@ -87,23 +92,23 @@ func (m *MockManager) AddHealthzCheck(arg0 string, arg1 healthz.Checker) error {
 }
 
 // AddHealthzCheck indicates an expected call of AddHealthzCheck.
-func (mr *MockManagerMockRecorder) AddHealthzCheck(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockManagerMockRecorder) AddHealthzCheck(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddHealthzCheck", reflect.TypeOf((*MockManager)(nil).AddHealthzCheck), arg0, arg1)
 }
 
-// AddMetricsExtraHandler mocks base method.
-func (m *MockManager) AddMetricsExtraHandler(arg0 string, arg1 http.Handler) error {
+// AddMetricsServerExtraHandler mocks base method.
+func (m *MockManager) AddMetricsServerExtraHandler(arg0 string, arg1 http.Handler) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AddMetricsExtraHandler", arg0, arg1)
+	ret := m.ctrl.Call(m, "AddMetricsServerExtraHandler", arg0, arg1)
 	ret0, _ := ret[0].(error)
 	return ret0
 }
 
-// AddMetricsExtraHandler indicates an expected call of AddMetricsExtraHandler.
-func (mr *MockManagerMockRecorder) AddMetricsExtraHandler(arg0, arg1 interface{}) *gomock.Call {
+// AddMetricsServerExtraHandler indicates an expected call of AddMetricsServerExtraHandler.
+func (mr *MockManagerMockRecorder) AddMetricsServerExtraHandler(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddMetricsExtraHandler", reflect.TypeOf((*MockManager)(nil).AddMetricsExtraHandler), arg0, arg1)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddMetricsServerExtraHandler", reflect.TypeOf((*MockManager)(nil).AddMetricsServerExtraHandler), arg0, arg1)
 }
 
 // AddReadyzCheck mocks base method.
@@ -115,7 +120,7 @@ func (m *MockManager) AddReadyzCheck(arg0 string, arg1 healthz.Checker) error {
 }
 
 // AddReadyzCheck indicates an expected call of AddReadyzCheck.
-func (mr *MockManagerMockRecorder) AddReadyzCheck(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockManagerMockRecorder) AddReadyzCheck(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddReadyzCheck", reflect.TypeOf((*MockManager)(nil).AddReadyzCheck), arg0, arg1)
 }
@@ -191,10 +196,10 @@ func (mr *MockManagerMockRecorder) GetConfig() *gomock.Call {
 }
 
 // GetControllerOptions mocks base method.
-func (m *MockManager) GetControllerOptions() v1alpha1.ControllerConfigurationSpec {
+func (m *MockManager) GetControllerOptions() config.Controller {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetControllerOptions")
-	ret0, _ := ret[0].(v1alpha1.ControllerConfigurationSpec)
+	ret0, _ := ret[0].(config.Controller)
 	return ret0
 }
 
@@ -213,7 +218,7 @@ func (m *MockManager) GetEventRecorderFor(arg0 string) record.EventRecorder {
 }
 
 // GetEventRecorderFor indicates an expected call of GetEventRecorderFor.
-func (mr *MockManagerMockRecorder) GetEventRecorderFor(arg0 interface{}) *gomock.Call {
+func (mr *MockManagerMockRecorder) GetEventRecorderFor(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetEventRecorderFor", reflect.TypeOf((*MockManager)(nil).GetEventRecorderFor), arg0)
 }
@@ -230,6 +235,20 @@ func (m *MockManager) GetFieldIndexer() client.FieldIndexer {
 func (mr *MockManagerMockRecorder) GetFieldIndexer() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetFieldIndexer", reflect.TypeOf((*MockManager)(nil).GetFieldIndexer))
+}
+
+// GetHTTPClient mocks base method.
+func (m *MockManager) GetHTTPClient() *http.Client {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "GetHTTPClient")
+	ret0, _ := ret[0].(*http.Client)
+	return ret0
+}
+
+// GetHTTPClient indicates an expected call of GetHTTPClient.
+func (mr *MockManagerMockRecorder) GetHTTPClient() *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetHTTPClient", reflect.TypeOf((*MockManager)(nil).GetHTTPClient))
 }
 
 // GetLogger mocks base method.
@@ -275,10 +294,10 @@ func (mr *MockManagerMockRecorder) GetScheme() *gomock.Call {
 }
 
 // GetWebhookServer mocks base method.
-func (m *MockManager) GetWebhookServer() *webhook.Server {
+func (m *MockManager) GetWebhookServer() webhook.Server {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetWebhookServer")
-	ret0, _ := ret[0].(*webhook.Server)
+	ret0, _ := ret[0].(webhook.Server)
 	return ret0
 }
 
@@ -286,20 +305,6 @@ func (m *MockManager) GetWebhookServer() *webhook.Server {
 func (mr *MockManagerMockRecorder) GetWebhookServer() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetWebhookServer", reflect.TypeOf((*MockManager)(nil).GetWebhookServer))
-}
-
-// SetFields mocks base method.
-func (m *MockManager) SetFields(arg0 interface{}) error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SetFields", arg0)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// SetFields indicates an expected call of SetFields.
-func (mr *MockManagerMockRecorder) SetFields(arg0 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetFields", reflect.TypeOf((*MockManager)(nil).SetFields), arg0)
 }
 
 // Start mocks base method.
@@ -311,7 +316,7 @@ func (m *MockManager) Start(arg0 context.Context) error {
 }
 
 // Start indicates an expected call of Start.
-func (mr *MockManagerMockRecorder) Start(arg0 interface{}) *gomock.Call {
+func (mr *MockManagerMockRecorder) Start(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Start", reflect.TypeOf((*MockManager)(nil).Start), arg0)
 }
