@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: controllers/providers/dns/dns.go
+//
+// Generated by this command:
+//
+//	mockgen -package=mocks -destination=controllers/mocks/provider_mock.go -source=controllers/providers/dns/dns.go Provider
+//
 
 // Package mocks is a generated GoMock package.
 package mocks
@@ -25,9 +30,10 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	v1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
 	assistant "github.com/k8gb-io/k8gb/controllers/providers/assistant"
+	gomock "go.uber.org/mock/gomock"
+	client "sigs.k8s.io/controller-runtime/pkg/client"
 	endpoint "sigs.k8s.io/external-dns/endpoint"
 )
 
@@ -63,23 +69,23 @@ func (m *MockProvider) CreateZoneDelegationForExternalDNS(arg0 *v1beta1.Gslb) er
 }
 
 // CreateZoneDelegationForExternalDNS indicates an expected call of CreateZoneDelegationForExternalDNS.
-func (mr *MockProviderMockRecorder) CreateZoneDelegationForExternalDNS(arg0 interface{}) *gomock.Call {
+func (mr *MockProviderMockRecorder) CreateZoneDelegationForExternalDNS(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateZoneDelegationForExternalDNS", reflect.TypeOf((*MockProvider)(nil).CreateZoneDelegationForExternalDNS), arg0)
 }
 
 // Finalize mocks base method.
-func (m *MockProvider) Finalize(arg0 *v1beta1.Gslb) error {
+func (m *MockProvider) Finalize(arg0 *v1beta1.Gslb, arg1 client.Client) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Finalize", arg0)
+	ret := m.ctrl.Call(m, "Finalize", arg0, arg1)
 	ret0, _ := ret[0].(error)
 	return ret0
 }
 
 // Finalize indicates an expected call of Finalize.
-func (mr *MockProviderMockRecorder) Finalize(arg0 interface{}) *gomock.Call {
+func (mr *MockProviderMockRecorder) Finalize(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Finalize", reflect.TypeOf((*MockProvider)(nil).Finalize), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Finalize", reflect.TypeOf((*MockProvider)(nil).Finalize), arg0, arg1)
 }
 
 // GetExternalTargets mocks base method.
@@ -91,24 +97,9 @@ func (m *MockProvider) GetExternalTargets(arg0 string) assistant.Targets {
 }
 
 // GetExternalTargets indicates an expected call of GetExternalTargets.
-func (mr *MockProviderMockRecorder) GetExternalTargets(arg0 interface{}) *gomock.Call {
+func (mr *MockProviderMockRecorder) GetExternalTargets(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetExternalTargets", reflect.TypeOf((*MockProvider)(nil).GetExternalTargets), arg0)
-}
-
-// GslbIngressExposedIPs mocks base method.
-func (m *MockProvider) GslbIngressExposedIPs(arg0 *v1beta1.Gslb) ([]string, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GslbIngressExposedIPs", arg0)
-	ret0, _ := ret[0].([]string)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// GslbIngressExposedIPs indicates an expected call of GslbIngressExposedIPs.
-func (mr *MockProviderMockRecorder) GslbIngressExposedIPs(arg0 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GslbIngressExposedIPs", reflect.TypeOf((*MockProvider)(nil).GslbIngressExposedIPs), arg0)
 }
 
 // SaveDNSEndpoint mocks base method.
@@ -120,7 +111,7 @@ func (m *MockProvider) SaveDNSEndpoint(arg0 *v1beta1.Gslb, arg1 *endpoint.DNSEnd
 }
 
 // SaveDNSEndpoint indicates an expected call of SaveDNSEndpoint.
-func (mr *MockProviderMockRecorder) SaveDNSEndpoint(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockProviderMockRecorder) SaveDNSEndpoint(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SaveDNSEndpoint", reflect.TypeOf((*MockProvider)(nil).SaveDNSEndpoint), arg0, arg1)
 }
