@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: controllers/depresolver/resolver.go
+//
+// Generated by this command:
+//
+//	mockgen -package=mocks -destination=controllers/mocks/resolver_mock.go -source=controllers/depresolver/resolver.go GslbResolver
+//
 
 // Package mocks is a generated GoMock package.
 package mocks
@@ -26,9 +31,9 @@ import (
 	context "context"
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	v1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
 	depresolver "github.com/k8gb-io/k8gb/controllers/depresolver"
+	gomock "go.uber.org/mock/gomock"
 	client "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -64,7 +69,7 @@ func (m *MockGslbResolver) ResolveGslbSpec(ctx context.Context, gslb *v1beta1.Gs
 }
 
 // ResolveGslbSpec indicates an expected call of ResolveGslbSpec.
-func (mr *MockGslbResolverMockRecorder) ResolveGslbSpec(ctx, gslb, client interface{}) *gomock.Call {
+func (mr *MockGslbResolverMockRecorder) ResolveGslbSpec(ctx, gslb, client any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ResolveGslbSpec", reflect.TypeOf((*MockGslbResolver)(nil).ResolveGslbSpec), ctx, gslb, client)
 }
