@@ -20,21 +20,22 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 
 import (
 	"fmt"
-	"io/ioutil"
+	"os"
 	"reflect"
 	"testing"
 
+	utils2 "github.com/k8gb-io/k8gb/controllers/utils"
+
 	k8gbv1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
 	"github.com/k8gb-io/k8gb/controllers/depresolver"
-	"github.com/k8gb-io/k8gb/controllers/internal/utils"
 	"github.com/k8gb-io/k8gb/controllers/mocks"
 
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 
-	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
+	"go.uber.org/mock/gomock"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	externaldns "sigs.k8s.io/external-dns/endpoint"
 
@@ -45,6 +46,11 @@ import (
 )
 
 // test data
+var targetIPs = []string{
+	"10.0.1.38",
+	"10.0.1.40",
+	"10.0.1.39",
+}
 var a = struct {
 	Config              depresolver.Config
 	Gslb                *k8gbv1beta1.Gslb
@@ -53,9 +59,10 @@ var a = struct {
 }{
 	Config: depresolver.Config{
 		ReconcileRequeueSeconds: 30,
+		NSRecordTTL:             30,
 		ClusterGeoTag:           "us",
 		ExtClustersGeoTags:      []string{"za", "eu"},
-		EdgeDNSServers: []utils.DNSServer{
+		EdgeDNSServers: []utils2.DNSServer{
 			{
 				Host: "dns.cloud.example.com",
 				Port: 53,
@@ -66,16 +73,13 @@ var a = struct {
 		K8gbNamespace: "k8gb",
 	},
 	Gslb: func() *k8gbv1beta1.Gslb {
-		var crSampleYaml = "../../../deploy/crds/k8gb.absa.oss_v1beta1_gslb_cr.yaml"
-		gslbYaml, _ := ioutil.ReadFile(crSampleYaml)
-		gslb, _ := utils.YamlToGslb(gslbYaml)
+		var crSampleYaml = "../../../deploy/crds/k8gb.absa.oss_v1beta1_gslb_cr_roundrobin_ingress.yaml"
+		gslbYaml, _ := os.ReadFile(crSampleYaml)
+		gslb, _ := utils2.YamlToGslb(gslbYaml)
+		gslb.Status.LoadBalancer.ExposedIPs = targetIPs
 		return gslb
 	}(),
-	TargetIPs: []string{
-		"10.0.1.38",
-		"10.0.1.40",
-		"10.0.1.39",
-	},
+	TargetIPs: targetIPs,
 	TargetNSNamesSorted: []string{
 		"gslb-ns-eu-cloud.example.com",
 		"gslb-ns-us-cloud.example.com",
@@ -93,13 +97,13 @@ var expectedDNSEndpoint = &externaldns.DNSEndpoint{
 		Endpoints: []*externaldns.Endpoint{
 			{
 				DNSName:    a.Config.DNSZone,
-				RecordTTL:  30,
+				RecordTTL:  externaldns.TTL(a.Config.NSRecordTTL),
 				RecordType: "NS",
 				Targets:    a.TargetNSNamesSorted,
 			},
 			{
 				DNSName:    "gslb-ns-us-cloud.example.com",
-				RecordTTL:  30,
+				RecordTTL:  externaldns.TTL(a.Config.NSRecordTTL),
 				RecordType: "A",
 				Targets:    a.TargetIPs,
 			},
@@ -113,7 +117,6 @@ func TestCreateZoneDelegationOnExternalDNS(t *testing.T) {
 	defer ctrl.Finish()
 	m := mocks.NewMockAssistant(ctrl)
 	p := NewExternalDNS(a.Config, m)
-	m.EXPECT().GslbIngressExposedIPs(a.Gslb).Return(a.TargetIPs, nil).Times(1)
 	m.EXPECT().SaveDNSEndpoint(a.Config.K8gbNamespace, gomock.Eq(expectedDNSEndpoint)).Return(nil).Times(1).
 		Do(func(ns string, ep *externaldns.DNSEndpoint) {
 			require.True(t, reflect.DeepEqual(ep, expectedDNSEndpoint))
