@@ -21,7 +21,7 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"os"
 	"path/filepath"
 	"strconv"
 	"strings"
@@ -37,23 +37,28 @@ import (
 	"github.com/gruntwork-io/terratest/modules/random"
 	"github.com/gruntwork-io/terratest/modules/shell"
 	"github.com/stretchr/testify/require"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 	corev1 "k8s.io/api/core/v1"
 	networkingv1 "k8s.io/api/networking/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 type Workflow struct {
-	error      error
-	namespace  string
-	cluster    string
-	k8sOptions *k8s.KubectlOptions
-	t          *testing.T
-	settings   struct {
-		ingressResourcePath string
-		gslbResourcePath    string
-		ingressName         string
-		digUsingUDP         bool
+	error                  error
+	namespace              string
+	cluster                string
+	k8sOptions             *k8s.KubectlOptions
+	k8sOptionsIstioIngress *k8s.KubectlOptions
+	t                      *testing.T
+	settings               struct {
+		ingressType                     IngressType
+		istioVirtualServiceResourcePath string
+		istioGatewayResourcePath        string
+		ingressResourcePath             string
+		gslbResourcePath                string
+		ingressName                     string
+		ingressPort                     int
+		digUsingUDP                     bool
 	}
 	state struct {
 		namespaceCreated bool
@@ -86,21 +91,56 @@ type TestAppResult struct {
 
 // InstanceStatus provides a simplified overview of the instance status
 type InstanceStatus struct {
-	Annotation       string   `json:"annotation"`
-	AppMessage       string   `json:"app-msg"`
-	AppRunning       bool     `json:"podinfo-running"`
-	AppReplicas      string   `json:"podinfo-replicas"`
-	LocalTargets     []string `json:"local-targets-ip"`
-	Ingresses        []string `json:"ingress-ip"`
-	Dig              []string `json:"dig-result"`
-	CoreDNS          string   `json:"coredns-ip"`
-	GslbHealthStatus string   `json:"gslb-status"`
-	Cluster          string   `json:"cluster"`
-	Namespace        string   `json:"namespace"`
-	Endpoint0DNSName string   `json:"ep0-dns-name"`
-	Endpoint0Targets string   `json:"ep0-dns-targets"`
-	Endpoint1DNSName string   `json:"ep1-dns-name"`
-	Endpoint1Targets string   `json:"ep1-dns-targets"`
+	i                     *Instance
+	Annotation            string   `json:"annotation"`
+	AppMessage            string   `json:"app-msg"`
+	AppRunning            bool     `json:"podinfo-running"`
+	AppReplicas           string   `json:"podinfo-replicas"`
+	LocalTargets          []string `json:"local-targets-ip"`
+	Ingresses             []string `json:"ingress-ip"`
+	Dig                   []string `json:"dig-result"`
+	CoreDNS               string   `json:"coredns-ip"`
+	GslbHealthStatus      string   `json:"gslb-status"`
+	Cluster               string   `json:"cluster"`
+	Namespace             string   `json:"namespace"`
+	EndpointLocalDNSName  string   `json:"ep0-dns-name"`
+	EndpointLocalTargets  string   `json:"ep0-dns-targets"`
+	EndpointGlobalDNSName string   `json:"ep1-dns-name"`
+	EndpointGlobalTargets string   `json:"ep1-dns-targets"`
+}
+
+type IngressType int
+
+const (
+	IngressEmbedded IngressType = iota
+	IngressReferenced
+	IngressIstio
+)
+
+const (
+	istioIngressNamespace           = "istio-ingress"
+	istioIngressLoadBalancerService = "istio-ingressgateway"
+	kubernetesIngressPort           = 80
+	istioIngressPort                = 8080
+)
+
+var IngressTypes = []IngressType{
+	IngressEmbedded,
+	IngressReferenced,
+	IngressIstio,
+}
+
+func (it IngressType) String() string {
+	switch it {
+	case IngressEmbedded:
+		return "ingress embedded"
+	case IngressReferenced:
+		return "ingress referenced"
+	case IngressIstio:
+		return "ingress istio"
+	default:
+		return "ingress unknown"
+	}
 }
 
 func NewWorkflow(t *testing.T, cluster string, port int) *Workflow {
@@ -115,9 +155,37 @@ func NewWorkflow(t *testing.T, cluster string, port int) *Workflow {
 	w.cluster = cluster
 	w.namespace = fmt.Sprintf("k8gb-test-%s", strings.ToLower(random.UniqueId()))
 	w.k8sOptions = k8s.NewKubectlOptions(cluster, "", w.namespace)
+	w.k8sOptionsIstioIngress = k8s.NewKubectlOptions(cluster, "", istioIngressNamespace)
 	w.t = t
 	w.state.gslb.port = port
 	w.error = err
+	return w
+}
+
+// Enrich enriches the workflow with the relevant resources depending on the ingressType
+func (w *Workflow) Enrich(basePath string, host string, ingressType IngressType) *Workflow {
+	gslbPath := ""
+	ingressPort := kubernetesIngressPort
+
+	switch ingressType {
+	case IngressEmbedded:
+		gslbPath = basePath + ".yaml"
+	case IngressReferenced:
+		gslbPath = basePath + "-ref-gslb.yaml"
+		ingressPath := basePath + "-ref-ingress.yaml"
+		w = w.WithIngress(ingressPath)
+	case IngressIstio:
+		gslbPath = basePath + "-istio-gslb.yaml"
+		virtualServicePath := basePath + "-istio-virtualservice.yaml"
+		gatewayPath := basePath + "-istio-gateway.yaml"
+		w = w.WithIstio(virtualServicePath, gatewayPath)
+		ingressPort = istioIngressPort
+	}
+
+	w = w.WithGslb(gslbPath, host)
+	w.settings.ingressType = ingressType
+	w.settings.ingressPort = ingressPort
+
 	return w
 }
 
@@ -125,7 +193,36 @@ func (w *Workflow) WithIngress(path string) *Workflow {
 	if path == "" {
 		w.error = fmt.Errorf("empty ingress resource path")
 	}
-	w.settings.ingressResourcePath = path
+	var err error
+	w.settings.ingressResourcePath, err = filepath.Abs(path)
+	if err != nil {
+		w.error = fmt.Errorf("reading %s; %s", path, err)
+	}
+	w.settings.ingressName, err = w.getManifestName(w.settings.ingressResourcePath)
+	if err != nil {
+		w.error = err
+	}
+	return w
+}
+
+func (w *Workflow) WithIstio(virtualServicePath, gatewayPath string) *Workflow {
+	if virtualServicePath == "" {
+		w.error = fmt.Errorf("empty istio virtual service resource path")
+	}
+	if gatewayPath == "" {
+		w.error = fmt.Errorf("empty istio gateway resource path")
+	}
+
+	var err error
+	w.settings.istioVirtualServiceResourcePath, err = filepath.Abs(virtualServicePath)
+	if err != nil {
+		w.error = fmt.Errorf("reading %s; %s", virtualServicePath, err)
+	}
+	w.settings.istioGatewayResourcePath, err = filepath.Abs(gatewayPath)
+	if err != nil {
+		w.error = fmt.Errorf("reading %s; %s", gatewayPath, err)
+	}
+
 	return w
 }
 
@@ -148,9 +245,6 @@ func (w *Workflow) WithGslb(path, host string) *Workflow {
 		w.error = err
 	}
 	w.state.gslb.host = host
-	if err != nil {
-		w.error = err
-	}
 	return w
 }
 
@@ -173,7 +267,14 @@ func (w *Workflow) Start() (*Instance, error) {
 
 	// namespace
 	w.t.Logf("Create namespace %s", w.namespace)
-	k8s.CreateNamespace(w.t, w.k8sOptions, w.namespace)
+	istioInjection := "disabled"
+	if w.settings.ingressType == IngressIstio {
+		istioInjection = "enabled"
+	}
+	k8s.CreateNamespaceWithMetadata(w.t, w.k8sOptions, metav1.ObjectMeta{
+		Name:   w.namespace,
+		Labels: map[string]string{"istio-injection": istioInjection},
+	})
 	w.state.namespaceCreated = true
 
 	// app
@@ -211,12 +312,28 @@ func (w *Workflow) Start() (*Instance, error) {
 
 	// gslb
 	if w.settings.gslbResourcePath != "" {
-		w.t.Logf("Create ingress %s from %s", w.state.gslb.name, w.settings.gslbResourcePath)
-		k8s.KubectlApply(w.t, w.k8sOptions, w.settings.gslbResourcePath)
-		k8s.WaitUntilIngressAvailable(w.t, w.k8sOptions, w.state.gslb.name, 100, 1*time.Second)
-		ingress := k8s.GetIngress(w.t, w.k8sOptions, w.state.gslb.name)
-		require.Equal(w.t, ingress.Name, w.state.gslb.name)
-		w.settings.ingressName = w.state.gslb.name
+		switch w.settings.ingressType {
+		case IngressEmbedded:
+			w.t.Logf("Create ingress %s from %s", w.state.gslb.name, w.settings.gslbResourcePath)
+			k8s.KubectlApply(w.t, w.k8sOptions, w.settings.gslbResourcePath)
+			k8s.WaitUntilIngressAvailable(w.t, w.k8sOptions, w.state.gslb.name, 100, 5*time.Second)
+			ingress := k8s.GetIngress(w.t, w.k8sOptions, w.state.gslb.name)
+			require.Equal(w.t, ingress.Name, w.state.gslb.name)
+			w.settings.ingressName = w.state.gslb.name
+		case IngressReferenced:
+			w.t.Logf("Create ingress %s from %s", w.settings.ingressName, w.settings.ingressResourcePath)
+			k8s.KubectlApply(w.t, w.k8sOptions, w.settings.ingressResourcePath)
+			k8s.WaitUntilIngressAvailable(w.t, w.k8sOptions, w.settings.ingressName, 100, 5*time.Second)
+			w.t.Logf("Create gslb %s from %s", w.state.gslb.name, w.settings.gslbResourcePath)
+			k8s.KubectlApply(w.t, w.k8sOptions, w.settings.gslbResourcePath)
+		case IngressIstio:
+			w.t.Logf("Create istio virtual service from %s", w.settings.istioVirtualServiceResourcePath)
+			k8s.KubectlApply(w.t, w.k8sOptions, w.settings.istioVirtualServiceResourcePath)
+			w.t.Logf("Create istio gateway from %s", w.settings.istioGatewayResourcePath)
+			k8s.KubectlApply(w.t, w.k8sOptionsIstioIngress, w.settings.istioGatewayResourcePath)
+			w.t.Logf("Create gslb %s from %s", w.state.gslb.name, w.settings.gslbResourcePath)
+			k8s.KubectlApply(w.t, w.k8sOptions, w.settings.gslbResourcePath)
+		}
 	}
 
 	return &Instance{
@@ -231,7 +348,7 @@ func (w *Workflow) getManifestName(path string) (string, error) {
 		} `yaml:"metadata"`
 	}{}
 
-	yamlFile, err := ioutil.ReadFile(path)
+	yamlFile, err := os.ReadFile(path)
 	if err != nil {
 		return "", fmt.Errorf("parse %s; %s", path, err)
 	}
@@ -247,6 +364,9 @@ func (i *Instance) Kill() {
 	if i.w.state.namespaceCreated {
 		k8s.DeleteNamespace(i.w.t, i.w.k8sOptions, i.w.namespace)
 	}
+	if i.w.settings.istioGatewayResourcePath != "" {
+		k8s.KubectlDelete(i.w.t, i.w.k8sOptionsIstioIngress, i.w.settings.istioGatewayResourcePath)
+	}
 }
 
 func (i *Instance) ReapplyIngress(path string) {
@@ -258,7 +378,7 @@ func (i *Instance) ReapplyIngress(path string) {
 	require.NoError(i.w.t, err)
 	k8s.KubectlApply(i.w.t, i.w.k8sOptions, i.w.settings.ingressResourcePath)
 	// modifying inner state.gslb.name and ingress.Name has nothing to do with reading these values dynamically afterwards.
-	k8s.WaitUntilIngressAvailable(i.w.t, i.w.k8sOptions, i.w.state.gslb.name, 60, 1*time.Second)
+	k8s.WaitUntilIngressAvailable(i.w.t, i.w.k8sOptions, i.w.state.gslb.name, 60, 5*time.Second)
 	ingress := k8s.GetIngress(i.w.t, i.w.k8sOptions, i.w.state.gslb.name)
 	require.Equal(i.w.t, ingress.Name, i.w.state.gslb.name)
 	i.w.settings.ingressName = i.w.state.gslb.name
@@ -279,9 +399,17 @@ func (i *Instance) GetCoreDNSIP() string {
 
 func (i *Instance) GetIngressIPs() []string {
 	var ingressIPs []string
-	ingress := k8s.GetIngress(i.w.t, i.w.k8sOptions, i.w.settings.ingressName)
-	for _, ip := range ingress.Status.LoadBalancer.Ingress {
-		ingressIPs = append(ingressIPs, ip.IP)
+	switch i.w.settings.ingressType {
+	case IngressIstio:
+		lbService := k8s.GetService(i.w.t, i.w.k8sOptionsIstioIngress, istioIngressLoadBalancerService)
+		for _, ip := range lbService.Status.LoadBalancer.Ingress {
+			ingressIPs = append(ingressIPs, ip.IP)
+		}
+	default:
+		ingress := k8s.GetIngress(i.w.t, i.w.k8sOptions, i.w.settings.ingressName)
+		for _, ip := range ingress.Status.LoadBalancer.Ingress {
+			ingressIPs = append(ingressIPs, ip.IP)
+		}
 	}
 	return ingressIPs
 }
@@ -327,7 +455,7 @@ func (i *Instance) WaitForLocalDNSEndpointExists() error {
 
 func (i *Instance) WaitForExternalDNSEndpointExists() error {
 	periodic := func() (result bool, err error) {
-		lep := i.Resources().GetExternalDNSEndpoint()
+		lep := i.Resources().GetK8gbExternalDNSEndpoint()
 		result = len(lep.Spec.Endpoints) > 0
 		return result, err
 	}
@@ -336,12 +464,12 @@ func (i *Instance) WaitForExternalDNSEndpointExists() error {
 
 func (r *Resources) WaitForExternalDNSEndpointHasTargets(epName string) error {
 	periodic := func() (result bool, err error) {
-		epx, err := r.GetExternalDNSEndpoint().GetEndpointByName(epName)
+		epx, err := r.GetK8gbExternalDNSEndpoint().GetEndpointByName(epName)
 		if err != nil {
 			return false, nil
 		}
 		result = len(epx.Targets) > 0
-		return result, err
+		return result, nil
 	}
 	return tickerWaiter(DefaultRetries, "ExternalDNSEndpoint has targets:", periodic)
 }
@@ -354,7 +482,7 @@ func (i *Instance) WaitForLocalDNSEndpointHasTargets(expectedIPs []string) error
 		result = EqualStringSlices(ep.Targets, expectedIPs)
 		return result, err
 	}
-	return tickerWaiter(DefaultRetries, "LocalDNSEndpoint targets:", periodic)
+	return tickerWaiter(160, "LocalDNSEndpoint targets:", periodic)
 }
 
 // WaitForExpected waits until GSLB dig doesn't return list of expected IP's
@@ -368,22 +496,26 @@ func (i *Instance) WaitForExpected(expectedIPs []string) (err error) {
 
 // WaitForAppIsRunning waits until app has one pod running
 func (i *Instance) WaitForAppIsRunning() (err error) {
-	return i.waitForApp(func(instances int) bool { return instances > 0 })
+	return i.waitForApp(func(instances int) bool { return instances > 0 }, false)
 }
 
 // WaitForAppIsStopped waits until app has 0 pods running
 func (i *Instance) WaitForAppIsStopped() (err error) {
-	return i.waitForApp(func(instances int) bool { return instances == 0 })
+	return i.waitForApp(func(instances int) bool { return instances == 0 }, true)
 }
 
 // WaitForAppIsRunning waits until app has one pod running
-func (i *Instance) waitForApp(predicate func(instances int) bool) (err error) {
+func (i *Instance) waitForApp(predicate func(instances int) bool, stop bool) (err error) {
 	const (
 		description = "Wait for app is running"
 		maxRetries  = 50
 		waitSeconds = 2
 	)
-
+	op := "stop"
+	if !stop {
+		op = "start"
+	}
+	// first condition is to have replicas
 	for n := 0; n < maxRetries; n++ {
 		var r int
 		rep := i.GetStatus("").AppReplicas
@@ -397,12 +529,49 @@ func (i *Instance) waitForApp(predicate func(instances int) bool) (err error) {
 		}
 		if predicate(r) {
 			i.w.t.Logf("%s found match: Expected:(%v)", description, r)
-			return nil
+			break
 		}
 		i.w.t.Logf("Application %s is not in expected state. Waiting for %d seconds...", i.w.state.testApp.name, waitSeconds)
 		time.Sleep(waitSeconds * time.Second)
 	}
-	return retry.MaxRetriesExceeded{Description: "Unable to start Podinfo app", MaxRetries: maxRetries}
+	i.w.t.Logf("Wait for ExternalDNSEndpoint %s.%s to be filled by targets %s", i.w.state.gslb.name, i.w.namespace, i.w.state.gslb.host)
+	// second conditions
+	for n := 0; n < maxRetries/2; n++ {
+		ep, err := i.Resources().GetExternalDNSEndpointByName(i.w.state.gslb.name, i.w.namespace).GetEndpointByName(fmt.Sprintf("localtargets-%s", i.w.state.gslb.host))
+		if err != nil {
+			// app is already stopped and cant be found
+			if stop && err.Error() == notFoundError {
+				i.w.t.Logf("App is stopped %s", i.w.state.testApp.name)
+				break
+			}
+			i.w.t.Logf("Error waiting for the app %s. %s", i.w.state.testApp.name, err)
+			require.NoError(i.w.t, err)
+		}
+		// waiting for the start
+		if !stop && len(ep.Targets) == 0 {
+			i.w.t.Logf("Waiting for %s to be started. Waiting for %d seconds...", i.w.state.testApp.name, waitSeconds)
+			time.Sleep(waitSeconds * time.Second)
+			continue
+		}
+		// waiting for the stop
+		if stop && len(ep.Targets) != 0 {
+			i.w.t.Logf("Waiting for %s to be stopped. Waiting for %d seconds...", i.w.state.testApp.name, waitSeconds)
+			time.Sleep(waitSeconds * time.Second)
+			continue
+		}
+	}
+	i.w.t.Logf("Wait for coreDNS to be filled by local targets %s", i.w.state.gslb.host)
+	for n := 0; n < maxRetries/2; n++ {
+		localTargets := i.GetLocalTargets()
+		if (!stop && len(localTargets) == 0) || (stop && len(localTargets) != 0) {
+			i.w.t.Logf("Waiting for coreDNS to be filled by local targets %s. Waiting for %d seconds...", i.w.state.gslb.host, waitSeconds)
+			time.Sleep(waitSeconds * time.Second)
+			continue
+		}
+		return nil
+	}
+
+	return retry.MaxRetriesExceeded{Description: "Unable to " + op + " Podinfo app", MaxRetries: maxRetries}
 }
 
 // String retrieves rough information about cluster
@@ -432,9 +601,14 @@ func (i *Instance) HitTestApp() (result *TestAppResult) {
 	var err error
 	result = new(TestAppResult)
 	coreDNSIP := i.GetCoreDNSIP()
-	command := []string{"sh", "-c", fmt.Sprintf("wget -qO - %s", i.w.state.gslb.host)}
-	for t := 0; t < 3; t++ {
+	command := []string{"sh", "-c", fmt.Sprintf("wget -qO - %s:%d", i.w.state.gslb.host, i.w.settings.ingressPort)}
+	for t := 0; t < 60; t++ {
 		result.Body, err = RunBusyBoxCommand(i.w.t, i.w.k8sOptions, coreDNSIP, command)
+		if strings.Contains(result.Body, "503") {
+			i.w.t.Log("podinfo returns 503, trying again....")
+			time.Sleep(time.Second * 1)
+			continue
+		}
 		require.NoError(i.w.t, err, "busybox", command, result.Body)
 		if strings.HasPrefix(result.Body, "{") {
 			break
@@ -478,25 +652,25 @@ func (i *Instance) GetStatus(annotation string) (s *InstanceStatus) {
 	if err != nil {
 		s.GslbHealthStatus = na
 	}
-	s.Endpoint0DNSName, err = k8s.RunKubectlAndGetOutputE(i.w.t, i.w.k8sOptions, "get", "dnsendpoints.externaldns.k8s.io", "test-gslb", "-o",
+	s.EndpointLocalDNSName, err = k8s.RunKubectlAndGetOutputE(i.w.t, i.w.k8sOptions, "get", "dnsendpoints.externaldns.k8s.io", "test-gslb", "-o",
 		"custom-columns=SERVICESTATUS:.spec.endpoints[0].dnsName", "--no-headers")
 	if err != nil {
-		s.Endpoint0DNSName = na
+		s.EndpointLocalDNSName = na
 	}
-	s.Endpoint0Targets, err = k8s.RunKubectlAndGetOutputE(i.w.t, i.w.k8sOptions, "get", "dnsendpoints.externaldns.k8s.io", "test-gslb", "-o",
+	s.EndpointLocalTargets, err = k8s.RunKubectlAndGetOutputE(i.w.t, i.w.k8sOptions, "get", "dnsendpoints.externaldns.k8s.io", "test-gslb", "-o",
 		"custom-columns=SERVICESTATUS:.spec.endpoints[0].targets", "--no-headers")
 	if err != nil {
-		s.Endpoint0Targets = na
+		s.EndpointLocalTargets = na
 	}
-	s.Endpoint1DNSName, err = k8s.RunKubectlAndGetOutputE(i.w.t, i.w.k8sOptions, "get", "dnsendpoints.externaldns.k8s.io", "test-gslb", "-o",
+	s.EndpointGlobalDNSName, err = k8s.RunKubectlAndGetOutputE(i.w.t, i.w.k8sOptions, "get", "dnsendpoints.externaldns.k8s.io", "test-gslb", "-o",
 		"custom-columns=SERVICESTATUS:.spec.endpoints[1].dnsName", "--no-headers")
 	if err != nil {
-		s.Endpoint1DNSName = na
+		s.EndpointGlobalDNSName = na
 	}
-	s.Endpoint1Targets, err = k8s.RunKubectlAndGetOutputE(i.w.t, i.w.k8sOptions, "get", "dnsendpoints.externaldns.k8s.io", "test-gslb", "-o",
+	s.EndpointGlobalTargets, err = k8s.RunKubectlAndGetOutputE(i.w.t, i.w.k8sOptions, "get", "dnsendpoints.externaldns.k8s.io", "test-gslb", "-o",
 		"custom-columns=SERVICESTATUS:.spec.endpoints[1].targets", "--no-headers")
 	if err != nil {
-		s.Endpoint1Targets = na
+		s.EndpointGlobalTargets = na
 	}
 	return
 }
@@ -510,14 +684,14 @@ func waitForLocalGSLBNew(t *testing.T, host string, port int, expectedResult []s
 		t,
 		"Wait for failover to happen and coredns to pickup new values...",
 		DefaultRetries,
-		time.Second*1,
+		time.Second*3,
 		func() ([]string, error) { return dns.Dig("localhost:"+strconv.Itoa(port), host, isUdp) },
 		expectedResult)
 }
 
 // tickerWaiter periodically executes periodic function
 func tickerWaiter(timeoutSeconds int, name string, periodic func() (result bool, err error)) error {
-	const interval = 2
+	const interval = 5
 	var cycles = timeoutSeconds / interval
 	t := 0
 	for range time.NewTicker(interval * time.Second).C {
@@ -554,14 +728,10 @@ type Resources struct {
 }
 
 // GslbSpecProperty returns actual value of one Spec property, e.g: `spec.ingress.rules[0].host`
-func (r *Resources) GslbSpecProperty(specPath string) string {
-	actualValue, _ := k8s.RunKubectlAndGetOutputE(r.i.w.t, r.i.w.k8sOptions, "get", "gslb", r.i.w.state.gslb.name,
+func (g *Gslb) GslbSpecProperty(specPath string) string {
+	actualValue, _ := k8s.RunKubectlAndGetOutputE(g.i.w.t, g.i.w.k8sOptions, "get", "gslb", g.i.w.state.gslb.name,
 		"-o", fmt.Sprintf("custom-columns=SERVICESTATUS:%s", specPath), "--no-headers")
 	return actualValue
-}
-
-func (r *Resources) Ingress() *networkingv1.Ingress {
-	return k8s.GetIngress(r.i.w.t, r.i.w.k8sOptions, r.i.w.settings.ingressName)
 }
 
 func (r *Resources) GetLocalDNSEndpoint() DNSEndpoint {
@@ -570,13 +740,17 @@ func (r *Resources) GetLocalDNSEndpoint() DNSEndpoint {
 	return ep
 }
 
-func (r *Resources) GetExternalDNSEndpoint() DNSEndpoint {
-	ep, err := r.getDNSEndpoint("k8gb-ns-extdns", "k8gb")
+func (r *Resources) GetK8gbExternalDNSEndpoint() DNSEndpoint {
+	return r.GetExternalDNSEndpointByName("k8gb-ns-extdns", "k8gb")
+}
+
+func (r *Resources) GetExternalDNSEndpointByName(name, namespace string) DNSEndpoint {
+	ep, err := r.getDNSEndpoint(name, namespace)
 	r.i.continueIfK8sResourceNotFound(err)
 	return ep
 }
 
-func (i *Instance) logIfError(err error, message string, args ...interface{}) {
+func (i *Instance) logIfError(err error, message string, args ...any) {
 	if err != nil {
 		i.w.t.Logf(message, args)
 	}
@@ -591,4 +765,59 @@ func (r *Resources) getDNSEndpoint(epName, ns string) (ep DNSEndpoint, err error
 	}
 	err = json.Unmarshal([]byte(j), &ep)
 	return ep, err
+}
+
+type Gslb struct {
+	i *Instance
+}
+
+type Ingress struct {
+	*networkingv1.Ingress
+	i *Instance
+}
+
+func (r *Resources) Ingress() *Ingress {
+	var ing *networkingv1.Ingress
+	ing = k8s.GetIngress(r.i.w.t, r.i.w.k8sOptions, r.i.w.settings.ingressName)
+	return &Ingress{
+		Ingress: ing,
+		i:       r.i,
+	}
+}
+
+func (r *Resources) Gslb() *Gslb {
+	return &Gslb{
+		i: r.i,
+	}
+}
+
+func (g *Gslb) GetAnnotations() (a map[string]string) {
+	m := struct {
+		Metadata struct {
+			Annotations map[string]string `yaml:"annotations"`
+		} `yaml:"metadata"`
+	}{}
+	strValue, err := k8s.RunKubectlAndGetOutputE(g.i.w.t, g.i.w.k8sOptions, "get", "gslb", g.i.w.state.gslb.name, "-ojson")
+	require.NoError(g.i.w.t, err)
+	err = json.Unmarshal([]byte(strValue), &m)
+	require.NoError(g.i.w.t, err)
+	return m.Metadata.Annotations
+}
+
+func (g *Gslb) PatchAnnotations(a map[string]string) (err error) {
+	return g.i.patchAnnotations(g.i.w.state.gslb.name, "gslb", a)
+}
+
+func (ing *Ingress) PatchAnnotations(a map[string]string) (err error) {
+	return ing.i.patchAnnotations(ing.GetName(), "ingress", a)
+}
+
+func (i *Instance) patchAnnotations(name, ktype string, a map[string]string) (err error) {
+	var data []string
+	for k, v := range a {
+		data = append(data, fmt.Sprintf(`"%s":"%s"`, k, v))
+	}
+	annotations := fmt.Sprintf("{\"metadata\":{\"annotations\":{%s}}}", strings.Join(data, ","))
+	_, err = k8s.RunKubectlAndGetOutputE(i.w.t, i.w.k8sOptions, "patch", ktype, name, "-p", annotations, "--type=merge")
+	return err
 }
