@@ -21,8 +21,8 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 
 import (
 	"fmt"
-	"io/ioutil"
 	"log"
+	"os"
 	"sort"
 	"strings"
 	"testing"
@@ -101,7 +101,7 @@ func DoWithRetryWaitingForValueE(t *testing.T, actionDescription string, maxRetr
 }
 
 func CreateGslb(t *testing.T, options *k8s.KubectlOptions, settings TestSettings, kubeResourcePath string) {
-	k8sManifestBytes, err := ioutil.ReadFile(kubeResourcePath)
+	k8sManifestBytes, err := os.ReadFile(kubeResourcePath)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -231,7 +231,7 @@ func WaitForLocalGSLB(t *testing.T, dnsServer string, dnsPort int, settings Test
 		t,
 		"Wait for failover to happen and coredns to pickup new values...",
 		300,
-		time.Second*1,
+		time.Second*3,
 		func() ([]string, error) { return Dig(t, dnsServer, dnsPort, host+settings.DNSZone, additionalArgs...) },
 		expectedResult)
 }
@@ -261,7 +261,7 @@ func RunBusyBoxCommand(t *testing.T, options *k8s.KubectlOptions, dns string, co
 	dnsOverride := fmt.Sprintf("{\"spec\":{\"dnsConfig\":{\"nameservers\":[\"%s\"]},\"dnsPolicy\": \"None\"}}", dns)
 	args := []string{}
 	kubectlCtx := []string{"--context", options.ContextName, "-n", options.Namespace}
-	containerArgs := []string{"run", "-i", "--rm", "busybox", "--restart", "Never", "--image", "busybox"}
+	containerArgs := []string{"run", "-i", "--rm", "busybox", "--restart", "Never", "--image", "busybox", "--labels", "sidecar.istio.io/inject=false"}
 	containerDNS := []string{"--overrides", dnsOverride}
 	appArgs := append([]string{"--"}, command...)
 	args = append(kubectlCtx, containerArgs...)
@@ -273,4 +273,18 @@ func RunBusyBoxCommand(t *testing.T, options *k8s.KubectlOptions, dns string, co
 		Env:     options.Env,
 	}
 	return shell.RunCommandAndGetOutputE(t, cmd)
+}
+
+// EqualAnnotations compares annotations of two maps
+// TODO: redundant to k8gb utils internal function.
+func EqualAnnotations(a, b map[string]string) bool {
+	if len(a) != len(b) {
+		return false
+	}
+	for k := range a {
+		if b[k] != a[k] {
+			return false
+		}
+	}
+	return true
 }
