@@ -1,5 +1,4 @@
 //go:build !ignore_autogenerated
-// +build !ignore_autogenerated
 
 /*
 Copyright 2022 The k8gb Contributors.
@@ -91,6 +90,7 @@ func (in *GslbSpec) DeepCopyInto(out *GslbSpec) {
 	*out = *in
 	in.Ingress.DeepCopyInto(&out.Ingress)
 	in.Strategy.DeepCopyInto(&out.Strategy)
+	in.ResourceRef.DeepCopyInto(&out.ResourceRef)
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new GslbSpec.
@@ -121,11 +121,24 @@ func (in *GslbStatus) DeepCopyInto(out *GslbStatus) {
 			if val == nil {
 				(*out)[key] = nil
 			} else {
-				in, out := &val, &outVal
+				inVal := (*in)[key]
+				in, out := &inVal, &outVal
 				*out = make([]string, len(*in))
 				copy(*out, *in)
 			}
 			(*out)[key] = outVal
+		}
+	}
+	in.LoadBalancer.DeepCopyInto(&out.LoadBalancer)
+	if in.Servers != nil {
+		in, out := &in.Servers, &out.Servers
+		*out = make([]*Server, len(*in))
+		for i := range *in {
+			if (*in)[i] != nil {
+				in, out := &(*in)[i], &(*out)[i]
+				*out = new(Server)
+				(*in).DeepCopyInto(*out)
+			}
 		}
 	}
 }
@@ -166,6 +179,83 @@ func (in *IngressSpec) DeepCopy() *IngressSpec {
 		return nil
 	}
 	out := new(IngressSpec)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *LoadBalancer) DeepCopyInto(out *LoadBalancer) {
+	*out = *in
+	if in.ExposedIPs != nil {
+		in, out := &in.ExposedIPs, &out.ExposedIPs
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new LoadBalancer.
+func (in *LoadBalancer) DeepCopy() *LoadBalancer {
+	if in == nil {
+		return nil
+	}
+	out := new(LoadBalancer)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *NamespacedName) DeepCopyInto(out *NamespacedName) {
+	*out = *in
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new NamespacedName.
+func (in *NamespacedName) DeepCopy() *NamespacedName {
+	if in == nil {
+		return nil
+	}
+	out := new(NamespacedName)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *ResourceRef) DeepCopyInto(out *ResourceRef) {
+	*out = *in
+	in.LabelSelector.DeepCopyInto(&out.LabelSelector)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ResourceRef.
+func (in *ResourceRef) DeepCopy() *ResourceRef {
+	if in == nil {
+		return nil
+	}
+	out := new(ResourceRef)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *Server) DeepCopyInto(out *Server) {
+	*out = *in
+	if in.Services != nil {
+		in, out := &in.Services, &out.Services
+		*out = make([]*NamespacedName, len(*in))
+		for i := range *in {
+			if (*in)[i] != nil {
+				in, out := &(*in)[i], &(*out)[i]
+				*out = new(NamespacedName)
+				**out = **in
+			}
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new Server.
+func (in *Server) DeepCopy() *Server {
+	if in == nil {
+		return nil
+	}
+	out := new(Server)
 	in.DeepCopyInto(out)
 	return out
 }
