@@ -21,18 +21,25 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	"context"
 	"fmt"
+	"reflect"
+
+	"github.com/k8gb-io/k8gb/controllers/utils"
 
 	"github.com/k8gb-io/k8gb/controllers/providers/metrics"
+	"github.com/k8gb-io/k8gb/controllers/refresolver"
+
+	"errors"
 
 	k8gbv1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
 	"github.com/k8gb-io/k8gb/controllers/depresolver"
-	"github.com/k8gb-io/k8gb/controllers/internal/utils"
 	"github.com/k8gb-io/k8gb/controllers/logging"
 	"github.com/k8gb-io/k8gb/controllers/providers/dns"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/trace"
-	"k8s.io/apimachinery/pkg/api/errors"
+	corev1 "k8s.io/api/core/v1"
+	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/client-go/tools/record"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -44,6 +51,7 @@ type GslbReconciler struct {
 	Config      *depresolver.Config
 	DepResolver depresolver.GslbResolver
 	DNSProvider dns.Provider
+	Recorder    record.EventRecorder
 	Tracer      trace.Tracer
 }
 
@@ -55,14 +63,17 @@ const (
 	splitBrainThresholdSecondsAnnotation = "k8gb.io/splitbrain-threshold-seconds"
 )
 
+var k8gbAnnotations = []string{strategyAnnotation, primaryGeoTagAnnotation, dnsTTLSecondsAnnotation, splitBrainThresholdSecondsAnnotation}
+
 var log = logging.Logger()
 
 var m = metrics.Metrics()
 
 // +kubebuilder:rbac:groups=k8gb.absa.oss,resources=gslbs,verbs=get;list;watch;create;update;patch;delete
 // +kubebuilder:rbac:groups=k8gb.absa.oss,resources=gslbs/status,verbs=get;update;patch
+// +kubebuilder:rbac:groups="",resources=events,verbs=create;patch
 
-// Reconcile runs main reconiliation loop
+// Reconcile runs main reconciliation loop
 func (r *GslbReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	ctx, span := r.Tracer.Start(ctx, "Reconcile")
 	defer span.End()
@@ -72,7 +83,7 @@ func (r *GslbReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.
 	gslb := &k8gbv1beta1.Gslb{}
 	err := r.Get(ctx, req.NamespacedName, gslb)
 	if err != nil {
-		if errors.IsNotFound(err) {
+		if k8serrors.IsNotFound(err) {
 			// Request object not found, could have been deleted after reconcile request.
 			// Owned objects are automatically garbage collected. For additional cleanup logic use finalizers.
 			// Return and don't requeue
@@ -89,6 +100,7 @@ func (r *GslbReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.
 	}
 	log.Debug().
 		Str("gslb", gslb.Name).
+		Str("namespace", gslb.Namespace).
 		Interface("strategy", gslb.Spec.Strategy).
 		Msg("Resolved strategy")
 	// == Finalizer business ==
@@ -137,17 +149,53 @@ func (r *GslbReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.
 	}
 
 	// == Ingress ==========
-	ingress, err := r.gslbIngress(gslb)
-	if err != nil {
-		m.IncrementError(gslb)
-		return result.RequeueError(err)
+	if reflect.DeepEqual(gslb.Spec.ResourceRef, k8gbv1beta1.ResourceRef{}) {
+		ingress, err := r.gslbIngress(gslb)
+		if err != nil {
+			m.IncrementError(gslb)
+			return result.RequeueError(err)
+		}
+
+		err = r.saveIngress(gslb, ingress)
+		if err != nil {
+			m.IncrementError(gslb)
+			return result.RequeueError(err)
+		}
 	}
 
-	err = r.saveIngress(gslb, ingress)
+	// == Reference resolution ==
+	refResolver, err := refresolver.New(gslb, r.Client)
+	if err != nil {
+		m.IncrementError(gslb)
+		errorMsg := fmt.Sprintf("error resolving references (%s)", err)
+		r.Recorder.Event(gslb, corev1.EventTypeWarning, "ReconcileError", errorMsg)
+		return result.RequeueError(errors.New(errorMsg))
+	}
+	servers, err := refResolver.GetServers()
+	if err != nil {
+		m.IncrementError(gslb)
+		return result.RequeueError(fmt.Errorf("getting GSLB servers (%s)", err))
+	}
+	gslb.Status.Servers = servers
+
+	loadBalancerExposedIPs, err := refResolver.GetGslbExposedIPs(gslb.Annotations, r.Config.EdgeDNSServers)
+	if err != nil {
+		m.IncrementError(gslb)
+		return result.RequeueError(fmt.Errorf("getting load balancer exposed IPs (%s)", err))
+	}
+	gslb.Status.LoadBalancer.ExposedIPs = loadBalancerExposedIPs
+
+	log.Debug().
+		Str("gslb", gslb.Name).
+		Msg("Resolved LoadBalancer and Server configuration referenced by Ingress")
+
+	// == health status of applications ==
+	serviceHealth, err := r.getServiceHealthStatus(gslb)
 	if err != nil {
 		m.IncrementError(gslb)
 		return result.RequeueError(err)
 	}
+	gslb.Status.ServiceHealth = serviceHealth
 
 	// == external-dns dnsendpoints CRs ==
 	dnsEndpoint, err := r.gslbDNSEndpoint(gslb)
