@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: controllers/providers/dns/infoblox-client.go
+//
+// Generated by this command:
+//
+//	mockgen -package=mocks -destination=controllers/mocks/infoblox-client_mock.go -source=controllers/providers/dns/infoblox-client.go InfobloxClient
+//
 
 // Package mocks is a generated GoMock package.
 package mocks
@@ -25,8 +30,8 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	ibclient "github.com/infobloxopen/infoblox-go-client"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockInfobloxClient is a mock of InfobloxClient interface.
