@@ -21,15 +21,15 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	"time"
 
-	k8gbv1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
+	corev1 "k8s.io/api/core/v1"
 	externaldns "sigs.k8s.io/external-dns/endpoint"
 )
 
 type Assistant interface {
+	// GetCoreDNSService returns the CoreDNS Service
+	GetCoreDNSService() (*corev1.Service, error)
 	// CoreDNSExposedIPs retrieves list of exposed IP by CoreDNS
 	CoreDNSExposedIPs() ([]string, error)
-	// GslbIngressExposedIPs retrieves list of IP's exposed by all GSLB ingresses
-	GslbIngressExposedIPs(gslb *k8gbv1beta1.Gslb) ([]string, error)
 	// GetExternalTargets retrieves slice of targets from external clusters
 	GetExternalTargets(host string, extClusterNsNames map[string]string) (targets Targets)
 	// SaveDNSEndpoint update DNS endpoint or create new one if doesnt exist
