@@ -21,20 +21,19 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	k8gbv1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
 	"github.com/k8gb-io/k8gb/controllers/providers/assistant"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	externaldns "sigs.k8s.io/external-dns/endpoint"
 )
 
 type Provider interface {
 	// CreateZoneDelegationForExternalDNS handles delegated zone in Edge DNS
 	CreateZoneDelegationForExternalDNS(*k8gbv1beta1.Gslb) error
-	// GslbIngressExposedIPs retrieves list of IP's exposed by all GSLB ingresses
-	GslbIngressExposedIPs(*k8gbv1beta1.Gslb) ([]string, error)
 	// GetExternalTargets retrieves list of external targets for specified host
 	GetExternalTargets(string) assistant.Targets
 	// SaveDNSEndpoint update DNS endpoint in gslb or create new one if doesn't exist
 	SaveDNSEndpoint(*k8gbv1beta1.Gslb, *externaldns.DNSEndpoint) error
 	// Finalize finalize gslb in k8gbNamespace
-	Finalize(*k8gbv1beta1.Gslb) error
+	Finalize(*k8gbv1beta1.Gslb, client.Client) error
 	// String see: Stringer interface
 	String() string
 }
