@@ -21,9 +21,9 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	"testing"
 
-	"github.com/k8gb-io/k8gb/controllers/depresolver"
-	"github.com/k8gb-io/k8gb/controllers/internal/utils"
+	"github.com/k8gb-io/k8gb/controllers/utils"
 
+	"github.com/k8gb-io/k8gb/controllers/depresolver"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"k8s.io/apimachinery/pkg/runtime"
