@@ -27,11 +27,12 @@ import (
 	"sync"
 	"time"
 
+	"github.com/k8gb-io/k8gb/controllers/utils"
+
 	externaldns "sigs.k8s.io/external-dns/endpoint"
 
 	k8gbv1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
 	"github.com/k8gb-io/k8gb/controllers/depresolver"
-	"github.com/k8gb-io/k8gb/controllers/internal/utils"
 	"github.com/prometheus/client_golang/prometheus"
 	crm "sigs.k8s.io/controller-runtime/pkg/metrics"
 )
