@@ -22,6 +22,7 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 */
 
 import (
+	"fmt"
 	"k8gbterratest/utils"
 	"testing"
 
@@ -32,22 +33,30 @@ import (
 // TestFailoverPlayground is equal to k8gb failover test running on local playground.
 // see: https://github.com/k8gb-io/k8gb/blob/master/docs/local.md#failover
 func TestFailoverPlayground(t *testing.T) {
-	t.Parallel()
-	const host = "playground-failover.cloud.example.com"
-	const gslbPath = "../examples/failover-playground.yaml"
+	for _, ingressType := range utils.IngressTypes {
+		const basePath = "../examples/failover-playground"
+		const host = "playground-failover.cloud.example.com"
+
+		workflowEU := utils.NewWorkflow(t, "k3d-test-gslb1", 5053)
+		workflowUS := utils.NewWorkflow(t, "k3d-test-gslb2", 5054)
+		abstractTestFailoverPlayground(t, ingressType.String(),
+			workflowEU.Enrich(basePath, host, ingressType),
+			workflowUS.Enrich(basePath, host, ingressType),
+		)
+	}
+}
+
+func abstractTestFailoverPlayground(t *testing.T, testPrefix string, workflowEU, workflowUS *utils.Workflow) {
 	const euGeoTag = "eu"
 	const usGeoTag = "us"
 
-	instanceEU, err := utils.NewWorkflow(t, "k3d-test-gslb1", 5053).
-		WithGslb(gslbPath, host).
-		WithTestApp(euGeoTag).
-		Start()
+	workflowEU = workflowEU.WithTestApp(euGeoTag)
+	instanceEU, err := workflowEU.Start()
 	require.NoError(t, err)
 	defer instanceEU.Kill()
-	instanceUS, err := utils.NewWorkflow(t, "k3d-test-gslb2", 5054).
-		WithGslb(gslbPath, host).
-		WithTestApp(usGeoTag).
-		Start()
+
+	workflowUS = workflowUS.WithTestApp(usGeoTag)
+	instanceUS, err := workflowUS.Start()
 	require.NoError(t, err)
 	defer instanceUS.Kill()
 
@@ -64,7 +73,7 @@ func TestFailoverPlayground(t *testing.T) {
 		assert.Equal(t, geoTag, httpResult.Message)
 	}
 
-	t.Run("failover on two concurrent clusters with TestApp running", func(t *testing.T) {
+	t.Run(fmt.Sprintf("%s failover on two concurrent clusters with TestApp running", testPrefix), func(t *testing.T) {
 		err = instanceEU.WaitForAppIsRunning()
 		require.NoError(t, err)
 		err = instanceUS.WaitForAppIsRunning()
@@ -74,13 +83,14 @@ func TestFailoverPlayground(t *testing.T) {
 	euLocalTargets := instanceEU.GetLocalTargets()
 	usLocalTargets := instanceUS.GetLocalTargets()
 
-	t.Run("stop podinfo on eu cluster", func(t *testing.T) {
+	t.Run(fmt.Sprintf("%s stop podinfo on eu cluster", testPrefix), func(t *testing.T) {
 		instanceEU.StopTestApp()
-		require.NoError(t, instanceEU.WaitForAppIsStopped())
+		err = instanceEU.WaitForAppIsStopped()
+		require.NoError(t, err, "Expecting Test App on EU cluster to be stopped")
 		actAndAssert(t.Name(), usGeoTag, usLocalTargets)
 	})
 
-	t.Run("start podinfo again on eu cluster", func(t *testing.T) {
+	t.Run(fmt.Sprintf("%s start podinfo again on eu cluster", testPrefix), func(t *testing.T) {
 		instanceEU.StartTestApp()
 		require.NoError(t, instanceEU.WaitForAppIsRunning())
 		actAndAssert(t.Name(), euGeoTag, euLocalTargets)
