@@ -22,6 +22,7 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 */
 
 import (
+	"fmt"
 	"k8gbterratest/utils"
 	"testing"
 
@@ -29,24 +30,31 @@ import (
 )
 
 func TestFullFailover(t *testing.T) {
-	t.Parallel()
-	const host = "terratest-failover.cloud.example.com"
-	const gslbPath = "../examples/failover.yaml"
+	for _, ingressType := range utils.IngressTypes {
+		const basePath = "../examples/failover"
+		const host = "terratest-failover.cloud.example.com"
 
-	instanceEU, err := utils.NewWorkflow(t, "k3d-test-gslb1", 5053).
-		WithGslb(gslbPath, host).
-		WithTestApp("eu").
-		Start()
+		workflowEU := utils.NewWorkflow(t, "k3d-test-gslb1", 5053)
+		workflowUS := utils.NewWorkflow(t, "k3d-test-gslb2", 5054)
+		abstractTestFullFailover(t, ingressType.String(),
+			workflowEU.Enrich(basePath, host, ingressType),
+			workflowUS.Enrich(basePath, host, ingressType),
+		)
+	}
+}
+
+func abstractTestFullFailover(t *testing.T, testPrefix string, workflowEU, workflowUS *utils.Workflow) {
+	workflowEU = workflowEU.WithTestApp("eu")
+	instanceEU, err := workflowEU.Start()
 	require.NoError(t, err)
 	defer instanceEU.Kill()
-	instanceUS, err := utils.NewWorkflow(t, "k3d-test-gslb2", 5054).
-		WithGslb(gslbPath, host).
-		WithTestApp("us").
-		Start()
+
+	workflowUS = workflowUS.WithTestApp("us")
+	instanceUS, err := workflowUS.Start()
 	require.NoError(t, err)
 	defer instanceUS.Kill()
 
-	t.Run("failover on two concurrent clusters with podinfo running", func(t *testing.T) {
+	t.Run(fmt.Sprintf("%s failover on two concurrent clusters with podinfo running", testPrefix), func(t *testing.T) {
 		err = instanceEU.WaitForAppIsRunning()
 		require.NoError(t, err)
 		err = instanceUS.WaitForAppIsRunning()
@@ -56,7 +64,7 @@ func TestFullFailover(t *testing.T) {
 	euLocalTargets := instanceEU.GetLocalTargets()
 	usLocalTargets := instanceUS.GetLocalTargets()
 
-	t.Run("kill podinfo on the second cluster", func(t *testing.T) {
+	t.Run(fmt.Sprintf("%s kill podinfo on the second cluster", testPrefix), func(t *testing.T) {
 		instanceUS.StopTestApp()
 		err = instanceUS.WaitForExpected(euLocalTargets)
 		require.NoError(t, err)
@@ -64,7 +72,7 @@ func TestFullFailover(t *testing.T) {
 		require.NoError(t, err)
 	})
 
-	t.Run("kill podinfo on the first cluster", func(t *testing.T) {
+	t.Run(fmt.Sprintf("%s kill podinfo on the first cluster", testPrefix), func(t *testing.T) {
 		instanceEU.StopTestApp()
 		err = instanceEU.WaitForExpected([]string{})
 		require.NoError(t, err)
@@ -72,7 +80,7 @@ func TestFullFailover(t *testing.T) {
 		require.NoError(t, err)
 	})
 
-	t.Run("start podinfo on the second cluster", func(t *testing.T) {
+	t.Run(fmt.Sprintf("%s start podinfo on the second cluster", testPrefix), func(t *testing.T) {
 		instanceUS.StartTestApp()
 		err = instanceUS.WaitForExpected(usLocalTargets)
 		require.NoError(t, err)
@@ -80,7 +88,7 @@ func TestFullFailover(t *testing.T) {
 		require.NoError(t, err)
 	})
 
-	t.Run("start podinfo on the first cluster", func(t *testing.T) {
+	t.Run(fmt.Sprintf("%s start podinfo on the first cluster", testPrefix), func(t *testing.T) {
 		instanceEU.StartTestApp()
 		err = instanceEU.WaitForExpected(euLocalTargets)
 		require.NoError(t, err)
