@@ -22,9 +22,23 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 */
 
 import (
+	"fmt"
+	"k8gbterratest/utils"
 	"testing"
 )
 
 func TestFullRoundRobin(t *testing.T) {
-	abstractTestFullRoundRobin(t, settings.ClustersNumber)
+	for _, ingressType := range utils.IngressTypes {
+		const basePath = "../examples/roundrobin2"
+		const host = "roundrobin-test.cloud.example.com"
+
+		workflows := []*utils.Workflow{}
+		for i := 0; i < settings.ClustersNumber; i += 1 {
+			workflow := utils.NewWorkflow(t, fmt.Sprintf("k3d-test-gslb%d", i+1), 5053+i)
+			workflow.Enrich(basePath, host, ingressType)
+			workflows = append(workflows, workflow)
+		}
+
+		abstractTestFullRoundRobin(t, ingressType.String(), workflows)
+	}
 }
